requirements.txt
requests~=2.22.0
PyYAML~=5.4.1
urllib3~=1.25.11